Refactored the Users app old "V1" folder into the new "rest_framework" structure
