+++ api/app/signals/apps/users/rest_framework/fields/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: MPL-2.0
+# Copyright (C) 2018 - 2022 Gemeente Amsterdam
