+++ api/app/signals/apps/users/rest_framework/serializers/__init__.py
@@ -1,0 +1,23 @@
+# SPDX-License-Identifier: MPL-2.0
+# Copyright (C) 2018 - 2022 Gemeente Amsterdam
+from signals.apps.users.rest_framework.serializers.permission import PermissionSerializer
+from signals.apps.users.rest_framework.serializers.profile import (
+    ProfileDetailSerializer,
+    ProfileListSerializer
+)
+from signals.apps.users.rest_framework.serializers.role import RoleSerializer
+from signals.apps.users.rest_framework.serializers.user import (
+    UserDetailHALSerializer,
+    UserListHALSerializer,
+    UserNameListSerializer
+)
+
+__all__ = [
+    'PermissionSerializer',
+    'ProfileListSerializer',
+    'ProfileDetailSerializer',
+    'RoleSerializer',
+    'UserDetailHALSerializer',
+    'UserListHALSerializer',
+    'UserNameListSerializer',
+]
