+++ api/app/signals/apps/users/rest_framework/serializers/profile.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: MPL-2.0
-# Copyright (C) 2019 - 2021 Gemeente Amsterdam
+# Copyright (C) 2019 - 2022 Gemeente Amsterdam
 from rest_framework import serializers
 
 from signals.apps.signals.models import Department
