+++ api/app/signals/apps/users/rest_framework/views/__init__.py
@@ -1,0 +1,17 @@
+# SPDX-License-Identifier: MPL-2.0
+# Copyright (C) 2019 - 2022 Gemeente Amsterdam
+from signals.apps.users.rest_framework.views.permission import PermissionViewSet
+from signals.apps.users.rest_framework.views.role import RoleViewSet
+from signals.apps.users.rest_framework.views.user import (
+    AutocompleteUsernameListView,
+    LoggedInUserView,
+    UserViewSet
+)
+
+__all__ = [
+    'PermissionViewSet',
+    'RoleViewSet',
+    'UserViewSet',
+    'LoggedInUserView',
+    'AutocompleteUsernameListView',
+]
