+++ api/app/signals/apps/users/rest_framework/views/role.py
@@ -1,12 +1,12 @@
 # SPDX-License-Identifier: MPL-2.0
-# Copyright (C) 2019 - 2021 Gemeente Amsterdam
+# Copyright (C) 2019 - 2022 Gemeente Amsterdam
 from datapunt_api.rest import DatapuntViewSetWritable
 from django.contrib.auth.models import Group
 from django.db.models.functions import Lower
 from rest_framework.permissions import DjangoModelPermissions
 
 from signals.apps.api.generics.permissions import SIAPermissions
-from signals.apps.users.v1.serializers import RoleSerializer
+from signals.apps.users.rest_framework.serializers import RoleSerializer
 from signals.auth.backend import JWTAuthBackend
