+++ api/app/signals/apps/users/tests/rest_framework/test_filters.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: MPL-2.0
-# Copyright (C) 2021 Vereniging van Nederlandse Gemeenten, Gemeente Amsterdam
+# Copyright (C) 2021 - 2022 Vereniging van Nederlandse Gemeenten, Gemeente Amsterdam
 from signals.apps.signals.factories import DepartmentFactory
 from signals.apps.users.factories import UserFactory
 from signals.test.utils import SignalsBaseApiTestCase
