+++ api/app/signals/apps/users/tests/rest_framework/test_users.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: MPL-2.0
-# Copyright (C) 2019 - 2021 Gemeente Amsterdam
+# Copyright (C) 2019 - 2022 Gemeente Amsterdam
 import unittest
 
 from django.contrib.auth.models import Group, Permission
