--- api/app/signals/apps/users/tests/v1/__init__.py
@@ -1,2 +1,0 @@
-# SPDX-License-Identifier: MPL-2.0
-# Copyright (C) 2018 - 2021 Gemeente Amsterdam
