--- api/app/signals/apps/users/v1/serializers/__init__.py
@@ -1,20 +1,0 @@
-# SPDX-License-Identifier: MPL-2.0
-# Copyright (C) 2018 - 2021 Gemeente Amsterdam
-from signals.apps.users.v1.serializers.permission import PermissionSerializer
-from signals.apps.users.v1.serializers.profile import ProfileDetailSerializer, ProfileListSerializer
-from signals.apps.users.v1.serializers.role import RoleSerializer
-from signals.apps.users.v1.serializers.user import (
-    UserDetailHALSerializer,
-    UserListHALSerializer,
-    UserNameListSerializer
-)
-
-__all__ = [
-    'PermissionSerializer',
-    'ProfileListSerializer',
-    'ProfileDetailSerializer',
-    'RoleSerializer',
-    'UserDetailHALSerializer',
-    'UserListHALSerializer',
-    'UserNameListSerializer',
-]
