--- api/app/signals/apps/users/v1/views/__init__.py
@@ -1,17 +1,0 @@
-# SPDX-License-Identifier: MPL-2.0
-# Copyright (C) 2019 - 2021 Gemeente Amsterdam
-from signals.apps.users.v1.views.permission import PermissionViewSet
-from signals.apps.users.v1.views.role import RoleViewSet
-from signals.apps.users.v1.views.user import (
-    AutocompleteUsernameListView,
-    LoggedInUserView,
-    UserViewSet
-)
-
-__all__ = [
-    'PermissionViewSet',
-    'RoleViewSet',
-    'UserViewSet',
-    'LoggedInUserView',
-    'AutocompleteUsernameListView',
-]
